+++ changelog.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Unreleased
 ### Added
+- added imports in sem/IO/__init__.py
 - module ```chunking_evaluate```: a module that will compute various metrics. Will replace the module ```chunking_fscore```
 - new section in manual: how to train a new model with SEM
 - explicit handling of annotated CoNLL files. Use attribute ```taggings``` for taggins such as POS and ```chunkings``` for BIO-formatted taggings such as chunking or NER: ```<file format="conll" fields="field1,field2,POS,NER" word_field="field1" taggings="POS" chunkings="NER" />```
 ### Changed
+- corrected some typos in __main__.py
+- renamed "KeyIO.py" to "keyIO.py"
+- in ```label_consistency```, use whole document if no sentence segmentation
 - corrected a bug when changing reference segmentation
 - SEM can now train wapiti models. Required several changes
 - corrected setup: when missing resource file is directory, use copy_tree